requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
Deprecated==1.2.14
fonttools==4.53.1
idna==3.8
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyarrow==17.0.0
pybaseball==2.2.7
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.9.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scipy==1.14.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
